backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
alembic
python-dotenv
email-validator
aiofiles
pytest
pytest-asyncio
httpx
pydantic-settings
python-multipart
PyJWT
python-jose==3.3.0
argon2-cffi
psycopg2-binary
passlib